search-page-2
Search results on the search page allow the user to select “currently reading”, “want to read”, or “read” to place the book in a certain shelf.
If a book is assigned to a shelf on the main page and that book appears on the search page, the correct shelf should be selected on the search page.
If that book's shelf is changed on the search page, that change should be reflected on the main page as well.
The option "None" should be selected if a book has not been assigned to a shelf.